Code for Converting from Raw Data to Usable Data
We collect a lot of raw data, and we want to convert it to a form our
annotators and coders can explore and analyze. This code is a first pass
at that transformation.

Specifically, we create lists of posts found in Facebook and YouTube
data and extract account-level information into a separate collection.