Added code for maintaining old Post IDs
In the original code, we just incremented a counter for each post id for
the UmdPostId field. This is brittle in that if we re-run the same
collection code on an updated collection, we are likely to get
inconsistent numeric IDs. To fix this, we now have a script that builds
a map of platform post IDs to numeric IDs, and we check that map before
assigning a new post ID to the UmdPostId field.